created common library with dtos

+++ CatService/src/main/java/com/samb1232/catservice/database/DBService.java
@@ -14,7 +14,7 @@ import com.samb1232.catservice.database.entities.User;
 import com.samb1232.catservice.database.repos.CatReactionRepository;
 import com.samb1232.catservice.database.repos.CatRepository;
 import com.samb1232.catservice.database.repos.UserRepository;
-import com.samb1232.catservice.dto.TGUser;
+import com.samb1232.common.dto.TGUser;
 
 import jakarta.persistence.EntityNotFoundException;
 
@@ -108,11 +108,11 @@ public class DBService {
         }
     }               
 
-    public long getLikesCount(Long catId) {
+    public int getLikesCount(Long catId) {
         return catReactionRepository.countLikesByCatId(catId);
     }
 
-    public long getDislikesCount(Long catId) {
+    public int getDislikesCount(Long catId) {
         return catReactionRepository.countDislikesByCatId(catId);
     }
